+++ quill/include/quill/backend/BackendOptions.h
@@ -173,5 +173,20 @@ struct BackendOptions
    * calls. Changing this value only affects the performance of the backend worker.
    */
   std::chrono::milliseconds rdtsc_resync_interval = std::chrono::milliseconds{500};
+
+  /**
+   * This option enables a check that verifies the log message contains only printable characters
+   * before forwarding it to the sinks. This adds an extra layer of safety by filtering out
+   * non-printable characters from the log file. Any non-printable characters are converted to their
+   * equivalent hex value.
+   *
+   * The check applies only when at least one argument in a log statement is of type string.
+   *
+   * You can customize this callback to define your own range of printable characters if needed.
+   *
+   * To disable this check, you can provide:
+   *   std::function<bool(char c)> check_printable_char = {}
+   */
+  std::function<bool(char c)> check_printable_char = [](char c) { return c >= ' ' && c <= '~'; };
 };
 } // namespace quill